Switched to using inheritance for db controllers
Switched it up from my pseudo-dynamic runtime typechecking magic that
was the original implementation, to a more easily readable and
maintainable inheritance-based approach.

I said at the beginning of this course that I would not want to use
inheritance... but I have now seen the light and I have been converted.

Thank you Olli for showing me the way.